vlasovsolver/vectorclass_sve: renamed header according to guidelines
